+++ saleor/graphql/core/mutations.py
@@ -20,7 +20,7 @@ from ...core.exceptions import PermissionDenied
 from ...core.permissions import AccountPermissions
 from ..decorators import staff_member_or_app_required
 from ..utils import get_nodes, resolve_global_ids_to_primary_keys
-from .types import Error, File, Upload
+from .types import File, Upload
 from .types.common import UploadError
 from .utils import from_global_id_or_error, snake_to_camel_case
 from .utils.error_codes import get_error_code_from_error
@@ -48,32 +48,44 @@ def get_error_fields(error_type_class, error_type_field, deprecation_reason=None
     return {error_type_field: error_field}
 
 
-def validation_error_to_error_type(validation_error: ValidationError) -> list:
+def validation_error_to_error_type(
+    validation_error: ValidationError, error_type_class
+) -> list:
     """Convert a ValidationError into a list of Error types."""
     err_list = []
+    error_class_fields = set(error_type_class._meta.fields.keys())
     if hasattr(validation_error, "error_dict"):
         # convert field errors
         for field, field_errors in validation_error.error_dict.items():
             field = None if field == NON_FIELD_ERRORS else snake_to_camel_case(field)
             for err in field_errors:
-                err_list.append(
-                    (
-                        Error(field=field, message=err.messages[0]),
-                        get_error_code_from_error(err),
-                        err.params,
-                    )
+                error = error_type_class(
+                    field=field,
+                    message=err.messages[0],
+                    code=get_error_code_from_error(err),
                 )
+                attach_error_params(error, err.params, error_class_fields)
+                err_list.append(error)
     else:
         # convert non-field errors
         for err in validation_error.error_list:
-            err_list.append(
-                (
-                    Error(message=err.messages[0]),
-                    get_error_code_from_error(err),
-                    err.params,
-                )
+            error = error_type_class(
+                message=err.messages[0],
+                code=get_error_code_from_error(err),
             )
+            attach_error_params(error, err.params, error_class_fields)
+            err_list.append(error)
     return err_list
+
+
+def attach_error_params(error, params: dict, error_class_fields: set):
+    if not params:
+        return {}
+    # If some of the params key overlap with error class fields
+    # attach param value to the error
+    error_fields_in_params = set(params.keys()) & error_class_fields
+    for error_field in error_fields_in_params:
+        setattr(error, error_field, params[error_field])
 
 
 class ModelMutationOptions(MutationOptions):
@@ -338,30 +350,15 @@ class BaseMutation(graphene.Mutation):
 
     @classmethod
     def handle_errors(cls, error: ValidationError, **extra):
-        errors = validation_error_to_error_type(error)
-        return cls.handle_typed_errors(errors, **extra)
+        error_list = validation_error_to_error_type(error, cls._meta.error_type_class)
+        return cls.handle_typed_errors(error_list, **extra)
 
     @classmethod
     def handle_typed_errors(cls, errors: list, **extra):
         """Return class instance with errors."""
-
-        typed_errors = []
-        error_class_fields = set(cls._meta.error_type_class._meta.fields.keys())
-        for e, code, params in errors:
-            error_instance = cls._meta.error_type_class(
-                field=e.field, message=e.message, code=code
-            )
-            if params:
-                # If some of the params key overlap with error class fields
-                # attach param value to the error
-                error_fields_in_params = set(params.keys()) & error_class_fields
-                for error_field in error_fields_in_params:
-                    setattr(error_instance, error_field, params[error_field])
-            typed_errors.append(error_instance)
-
         if cls._meta.error_type_field is not None:
-            extra.update({cls._meta.error_type_field: typed_errors})
-        return cls(errors=typed_errors, **extra)
+            extra.update({cls._meta.error_type_field: errors})
+        return cls(errors=errors, **extra)
 
 
 class ModelMutation(BaseMutation):